more refactoring of table body and header
